+++ genai-web/packages/web/src/components/ui/mobile-menu/MobileMenu.tsx
@@ -128,6 +128,9 @@ export const MobileMenu = forwardRef<HTMLDialogElement, Props>((props, ref) => {
                     </li>
                   )}
                   <li>
+                    <MobileMenuItemLink label='ナレッジ管理' to='/knowledge' />
+                  </li>
+                  <li>
                     <MobileMenuItemLink label='利用履歴' to='/history' />
                   </li>
                   <li>